Move Api specs to their logical place
